refactor(activemodel): converge attribute_method?, errors assertions, dirty extras (RFC 0115)
- attribute-methods.ts: `isAttributeMethod` now makes both calls Rails'
  `attribute_method?` makes (attribute_methods.rb:541-543) —
  `respond_to_without_attributes?(:attributes) && attributes.include?(attr_name)`.
  ActiveRecord already overrides the instance predicate
  (activerecord/attribute_methods.rb:499-501), so the accessed-fields
  consequence stays confined.
- model.ts: delete the invented instance `typeForAttribute` wrapper (Rails
  defines `type_for_attribute` only on the class,
  attribute_registration.rb:43) and its wrong `attribute_for_inspect`
  Mirrors line; the `.type_for_attribute` tests now call the class method.
- errors.test.ts: converge five drifted bodies onto Rails' — same `Person`
  subject (errors_test.rb:6-30), same assertions, same literal values.
- dirty.test.ts: move the 36 TS-only extras into dirty.trails.test.ts.
